move tests to helper module and remove unnecessary trait for helper
